checking on precedent results and resutls on z
Check if the former results where eigenvalue weren't inside bounds were caused by theoritical or numerical errors. Comparisons with the results with property 1 were the solutions to check. Results confirm that it apparently comes from numerical errors.
REsults on z in order to know if z can be used in order to know which values of x are active or inactive