work to check the bounds (finished)
Eigenvalues weren(t inside the theoritical bounds with the precedent way to calculate the bounds. Now, they are: the system is repaired.
Th errors came from two sources. Firstly, how x and z are defined inside the formulation. To calculat the right bounds, you have to use o.x1, o.x2, o.z1 and o.z2, and not o.x and o.z. Secondly, a scale effect is present inside the formulation. So, use the input d1 and d2 is wrond, you have to use the o.d1 and o.d2.